2659 - disallow dynamic arrays in recipes
The rule is: every 'local' variable in a recipe must have a fixed size.
Arrays can only be directly used in a recipe if their type includes a
size. But we haven't been warning about this, and attempts to use array
variables could cause silent memory corruption.

(Hopefully this is the last hole in our type system.)