2789
Why do earlier unit tests start to fail if I load chessboard.mu fail
after correcting the definition of new-channel to be properly generic? I
have narrowed it down to a manual test but am still struggling to turn
it into a scenario.

It looks like the 'new' instruction is the culprit:
  first transform:
    load a recipe containing a call to new with a generic type
      parse_type_tree is called, and pollutes the Type_ordinal table with the generic type
  second transform:
    load a generic container containing the same generic type ingredient
    it fails to be detected as a generic type ingredient

To reproduce, compare these two commands:
  mu test
  mu test z.mu