2977 - draw new sandbox menu in edit/
Doesn't work as advertised yet. This is just the render piece, and
fixing all the tests.

I've been careful to try to break tests for edit once I implement the
button. Delete I can't ensure will break afterwards. Remember to test
clicking on multiple places on the menu.

Managing the screens is starting to grow onerous; maybe we need
something called normalize which clears some things. But the sandbox
menu can be on arbitrary lines..